+++ packages/cargo/jns42-core/src/schema_transforms/flatten.rs
@@ -4,7 +4,7 @@ macro_rules! generate_mod {
       use crate::models::SchemaArena;
       use std::iter;
 
-      pub fn transform(arena: &mut SchemaArena, key: u32) {
+      pub fn transform(arena: &mut SchemaArena, key: usize) {
         let item = arena.get_item(key);
 
         let mut item_new = item.clone();
